requirements.txt
python-telegram-bot>=20.0,<21.0
requests>=2.31.0
colorama>=0.4.6